requirements.txt
selenium==4.17.2
pytest==8.0.1
allure-pytest==2.13.2
pytest-xdist==3.5.0
pytest-rerunfailures==13.0
requests==2.31.0
Faker==19.6.1
answers~=0.0.4
eyes-selenium==5.18.0
setuptools==69.1.1
pytest-playwright==0.4.4
